[BEAM-1218] Move the names and auth files to appropriate directories
