+++ sdks/python/apache_beam/internal/gcp/auth.py
