+++ sdks/python/apache_beam/runners/dataflow/internal/dependency.py
@@ -64,7 +64,7 @@ import tempfile
 from apache_beam import utils
 from apache_beam import version as beam_version
 from apache_beam.internal import pickler
-from apache_beam.utils import names
+from apache_beam.runners.dataflow.internal import names
 from apache_beam.utils import processes
 from apache_beam.utils.pipeline_options import GoogleCloudOptions
 from apache_beam.utils.pipeline_options import SetupOptions
