+++ sdks/python/apache_beam/runners/dataflow/internal/dependency_test.py
@@ -25,7 +25,7 @@ import unittest
 
 from apache_beam import utils
 from apache_beam.runners.dataflow.internal import dependency
-from apache_beam.utils import names
+from apache_beam.runners.dataflow.internal import names
 from apache_beam.utils.pipeline_options import GoogleCloudOptions
 from apache_beam.utils.pipeline_options import PipelineOptions
 from apache_beam.utils.pipeline_options import SetupOptions
