+++ sdks/python/generate_pydoc.sh
@@ -26,6 +26,9 @@
 set -e
 
 # Create docs directory if it does not exist
+mkdir -p target/docs
+rm -rf target/docs/*
+
 mkdir -p target/docs/source
 
 # Exclude autogenerated API message definition files that aren't part of SDK.
